% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grounding.R
\name{ducksemantics_bebel_runner}
\alias{ducksemantics_bebel_runner}
\title{Create a BebeLM prompt provider}
\usage{
ducksemantics_bebel_runner(agent, on_event = NULL)
}
\arguments{
\item{agent}{An \code{Rbebelm} agent.}

\item{on_event}{Optional event handler.}
}
\value{
An object implementing \link{DucksemanticsPromptRunner}.
}
\description{
Create a BebeLM prompt provider
}
