% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducksemantics.R
\name{ducksemantics_cache_file}
\alias{ducksemantics_cache_file}
\title{Cache a source file}
\usage{
ducksemantics_cache_file(
  url,
  filename = basename(url),
  cache_dir = tools::R_user_dir("ducksemantics", "cache"),
  refresh = FALSE
)
}
\arguments{
\item{url}{Source URL.}

\item{filename}{Cache filename.}

\item{cache_dir}{Cache directory.}

\item{refresh}{Download even when the cache file already exists?}
}
\value{
Normalized path to the cached file.
}
\description{
Cache a source file
}
