% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ducksemantics_index_aliases}
\alias{ducksemantics_index_aliases}
\title{Build the lexical alias index}
\usage{
ducksemantics_index_aliases(conn, prefix = "semantic")
}
\arguments{
\item{conn}{DBI connection.}

\item{prefix}{Table prefix.}
}
\value{
Invisibly, index table name.
}
\description{
Build the lexical alias index
}
