% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ducksemantics_normalize}
\alias{ducksemantics_normalize}
\title{Normalize semantic text}
\usage{
ducksemantics_normalize(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
Normalized character vector.
}
\description{
Normalize semantic text
}
