% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ducksemantics_read_obo}
\alias{ducksemantics_read_obo}
\title{Read an OBO ontology into graph relations}
\usage{
ducksemantics_read_obo(
  path,
  family,
  source = basename(path),
  include_obsolete = FALSE
)
}
\arguments{
\item{path}{OBO file path.}

\item{family}{Ontology family.}

\item{source}{Alias source identity.}

\item{include_obsolete}{Include obsolete terms.}
}
\value{
A list of \code{nodes}, \code{aliases}, and \code{edges} data frames.
}
\description{
Read an OBO ontology into graph relations
}
