% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ducksemantics_schema_sql}
\alias{ducksemantics_schema_sql}
\title{DuckDB semantic graph schema}
\usage{
ducksemantics_schema_sql(prefix = "semantic")
}
\arguments{
\item{prefix}{Prefix for semantic tables.}
}
\value{
Character DDL statements.
}
\description{
DuckDB semantic graph schema
}
