% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ducksemantics_tables}
\alias{ducksemantics_tables}
\title{Semantic table names}
\usage{
ducksemantics_tables(prefix = "semantic")
}
\arguments{
\item{prefix}{Table prefix.}
}
\value{
A named character vector.
}
\description{
Semantic table names
}
