% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{DucksemanticsDbConnection}
\alias{DucksemanticsDbConnection}
\title{DBI connection reference}
\usage{
DucksemanticsDbConnection(value = NULL)
}
\arguments{
\item{value}{DBI connection.}
}
\value{
A \code{DucksemanticsDbConnection} object.
}
\description{
S7 value object for a valid DBI connection.
}
