% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{DucksemanticsScalarText}
\alias{DucksemanticsScalarText}
\title{Non-empty scalar text}
\usage{
DucksemanticsScalarText(value = character(0))
}
\arguments{
\item{value}{Character scalar.}
}
\value{
A \code{DucksemanticsScalarText} object.
}
\description{
S7 value object for a required non-empty character scalar.
}
