% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{DucksemanticsTable}
\alias{DucksemanticsTable}
\title{Data frame contract}
\arguments{
\item{value}{Data frame.}

\item{required}{Required column names.}

\item{allow_empty}{Whether zero-row input is allowed.}
}
\value{
A \code{DucksemanticsTable} object.
}
\description{
S7 value object for a data frame with required columns.
}
