% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{DucksemanticsTokenEmbeddingProvider}
\alias{DucksemanticsTokenEmbeddingProvider}
\title{Token-embedding-provider protocol}
\usage{
DucksemanticsTokenEmbeddingProvider
}
\description{
Token-embedding-provider protocol
}
