% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{fast_hpo_cr_available}
\alias{fast_hpo_cr_available}
\title{Test whether the FastHPOCR Python module can be imported}
\usage{
fast_hpo_cr_available(error = FALSE)
}
\arguments{
\item{error}{If \code{TRUE}, throw an informative error when the Python dependency
set is unavailable.}
}
\value{
\code{TRUE} if \code{FastHPOCR.HPOAnnotator} imports successfully, otherwise
\code{FALSE}.
}
\description{
Test whether the FastHPOCR Python module can be imported
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
fast_hpo_cr_available()
\dontshow{\}) # examplesIf}
}
