% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{fast_hpo_cr_config}
\alias{fast_hpo_cr_config}
\title{Show reticulate's active Python configuration}
\usage{
fast_hpo_cr_config()
}
\value{
The object returned by \code{\link[reticulate:py_config]{reticulate::py_config()}}.
}
\description{
Show reticulate's active Python configuration
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
if (fast_hpo_cr_available()) {
  fast_hpo_cr_config()
}
\dontshow{\}) # examplesIf}
}
