% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RfastHPOCR-package.R
\docType{package}
\name{RfastHPOCR-package}
\alias{RfastHPOCR}
\alias{RfastHPOCR-package}
\title{R bindings to \code{FastHPOCR}}
\description{
\code{RfastHPOCR} provides a small R interface to the upstream Python
\href{https://github.com/tudorgroza/fast_hpo_cr}{\code{FastHPOCR}} package using
\link[reticulate:reticulate]{reticulate}. It exposes helpers for installing the
Python dependency set, building FastHPOCR ontology indexes, creating
annotators, and converting annotation objects to R data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/RfastHPOCR}
  \item \url{https://github.com/tudorgroza/fast_hpo_cr}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/RfastHPOCR/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Other contributors:
\itemize{
  \item Tudor Groza (Author and copyright holder of the upstream 'FastHPOCR' Python package) [contributor, copyright holder]
}

}
\keyword{internal}
