% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducksemantics-package.R
\docType{package}
\name{ducksemantics-package}
\alias{ducksemantics}
\alias{ducksemantics-package}
\title{DuckDB-native semantic and retrieval relations}
\description{
\code{ducksemantics} owns ontology graphs, validated HPO observations, typed
release-cataloged Monarch relations, provider/snapshot-bound literature
retrieval, source-grounded judgments, and genuine model/provider protocols.
Bulk values are ordinary relations; evaluation belongs to \code{VariantStoryBench}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/ducksemantics}
  \item \url{https://sounkou-bioinfo.github.io/ducksemantics/}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/ducksemantics/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Authors:
\itemize{
  \item Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}
}

Other contributors:
\itemize{
  \item Human Phenotype Ontology (Bundled README fixture (CC BY 4.0)) [copyright holder]
}

}
\keyword{internal}
