% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducksemantics-package.R
\docType{package}
\name{ducksemantics-package}
\alias{ducksemantics}
\alias{ducksemantics-package}
\title{DuckDB-native semantic graph grounding}
\description{
\code{ducksemantics} owns a small graph-as-SQL substrate for semantic grounding:
graph tables, alias indexes, mention extraction, transitive closure, and
model judgment prompts over deterministic candidates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/ducksemantics}
  \item \url{https://sounkou-bioinfo.github.io/ducksemantics}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/ducksemantics/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

}
\keyword{internal}
