% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducksemantics.R
\name{ducksemantics_embedding_search}
\alias{ducksemantics_embedding_search}
\title{Search embeddings with DuckDB vector functions}
\usage{
ducksemantics_embedding_search(query, conn, prefix = "semantic")
}
\arguments{
\item{query}{A \link{DucksemanticsEmbeddingQuery} object from
\code{\link[=ducksemantics_embedding_query]{ducksemantics_embedding_query()}}.}

\item{conn}{DBI connection.}

\item{prefix}{Prefix used for semantic tables.}
}
\value{
Data frame ordered by best match.
}
\description{
Search embeddings with DuckDB vector functions
}
