% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducksemantics.R
\name{ducksemantics_init}
\alias{ducksemantics_init}
\title{Initialize semantic graph tables}
\usage{
ducksemantics_init(conn, prefix = "semantic")
}
\arguments{
\item{conn}{DBI connection.}

\item{prefix}{Prefix used for semantic tables.}
}
\value{
Invisibly, \code{conn}.
}
\description{
Initialize semantic graph tables
}
