% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{ducksemantics_json_judgment_parser}
\alias{ducksemantics_json_judgment_parser}
\title{Create the default JSON judgment parser}
\usage{
ducksemantics_json_judgment_parser()
}
\value{
An object implementing \link{DucksemanticsJudgmentParser}.
}
\description{
Create the default JSON judgment parser
}
