% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{ducksemantics_lexical_annotator}
\alias{ducksemantics_lexical_annotator}
\title{Create the default DuckDB lexical annotator}
\usage{
ducksemantics_lexical_annotator()
}
\value{
An object implementing \link{DucksemanticsAnnotator}.
}
\description{
Create the default DuckDB lexical annotator
}
