% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducksemantics.R
\name{ducksemantics_normalize}
\alias{ducksemantics_normalize}
\title{Normalize text for semantic grounding}
\usage{
ducksemantics_normalize(text)
}
\arguments{
\item{text}{Character vector.}
}
\value{
Normalized character vector.
}
\description{
Normalize text for semantic grounding
}
