% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducksemantics.R
\name{ducksemantics_record_judgments}
\alias{ducksemantics_record_judgments}
\title{Record semantic judgments}
\usage{
ducksemantics_record_judgments(conn, judgments, prefix = "semantic")
}
\arguments{
\item{conn}{DBI connection.}

\item{judgments}{Judgment data frame.}

\item{prefix}{Prefix used for semantic tables.}
}
\value{
Invisibly, \code{judgments}.
}
\description{
Record semantic judgments
}
