% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducksemantics.R
\name{ducksemantics_tables}
\alias{ducksemantics_tables}
\title{Semantic graph table names}
\usage{
ducksemantics_tables(prefix = "semantic")
}
\arguments{
\item{prefix}{Prefix used for generated table names.}
}
\value{
A named character vector.
}
\description{
Semantic graph table names
}
